+++ animation_nodes/nodes/matrix/distribute_matrices.pyx
@@ -5,7 +5,7 @@ from libc.math cimport sin, cos
 from ... utils.limits cimport INT_MAX
 from ... events import propertyChanged
 from ... base_types import AnimationNode
-from ... data_structures cimport Mesh, Matrix4x4List, Vector3DList, Spline
+from ... data_structures cimport Mesh, Matrix4x4List, Vector3DList, Spline, Interpolation
 from ... algorithms.rotations.rotation_and_direction cimport directionToMatrix_LowLevel
 from ... math cimport (Matrix4, Vector3, setTranslationMatrix,
     setMatrixTranslation, setRotationZMatrix, toVector3, scaleMatrix3x3Part)
@@ -37,10 +37,19 @@ splineDistributionMethodItems = (
     ("VERTICES", "Vertices", "", "NONE", 3),
 )
 
+searchItems = {
+    "Distribute Linear" : "LINEAR",
+    "Distribute Grid" : "GRID",
+    "Distribute Circle" : "CIRCLE",
+    "Distribute MESH" : "MESH",
+    "Distribute Spiral" : "SPIRAL",
+    "Distribute Spline" : "SPLINE",
+}
+
 class DistributeMatricesNode(bpy.types.Node, AnimationNode):
     bl_idname = "an_DistributeMatricesNode"
     bl_label = "Distribute Matrices"
-    bl_width_default = 160
+    searchTags = [(name, {"mode" : repr(op)}) for name, op in searchItems.items()]
 
     __annotations__ = {}
 
@@ -69,6 +78,10 @@ class DistributeMatricesNode(bpy.types.Node, AnimationNode):
     __annotations__["splineResolution"] = IntProperty(name = "Spline Resolution", min = 2, default = 20,
         description = "Increase to have a more accurate evaluation if the type is set to Uniform",
         update = propertyChanged)
+    
+    __annotations__["centerSpiral"] =  BoolProperty(name = "Center Spiral",
+        description = "Center the spiral along Z axis",
+        default = False, update = propertyChanged)
 
     def create(self):
         if self.mode == "LINEAR":
@@ -103,6 +116,9 @@ class DistributeMatricesNode(bpy.types.Node, AnimationNode):
             self.newInput("Float", "End Size", "endSize", value = 0.5, minValue = 0)
             self.newInput("Float", "Start Angle", "startAngle", value = 0)
             self.newInput("Float", "End Angle", "endAngel", value = 6 * PI)
+            self.newInput("Float", "Height", "spiralHeight", value = 0)
+            self.newInput("Interpolation", "Radius Interpolation", "radiusInterpolation", defaultDrawType = "PROPERTY_ONLY", hide = True)
+            self.newInput("Interpolation", "Height Interpolation", "heightInterpolation", defaultDrawType = "PROPERTY_ONLY", hide = True)
         elif self.mode == "SPLINE":
             self.newInput("Spline", "Spline", "spline", defaultDrawType = "PROPERTY_ONLY")
             if self.splineDistributionMethod == "STEP":
@@ -130,6 +146,8 @@ class DistributeMatricesNode(bpy.types.Node, AnimationNode):
             row.prop(self, "centerAlongZ", text = "Z", toggle = True)
         if self.mode == "SPLINE":
             col.prop(self, "splineDistributionMethod", text = "")
+        if self.mode == "SPIRAL":
+            layout.prop(self, "centerSpiral", toggle = True)
 
     def drawAdvanced(self, layout):
         if self.mode == "CIRCLE":
@@ -157,7 +175,8 @@ class DistributeMatricesNode(bpy.types.Node, AnimationNode):
             elif self.meshMode == "POLYGONS":
                 yield "matrices = self.execute_Polygons(mesh)"
         elif self.mode == "SPIRAL":
-            yield "matrices = self.execute_Spiral(amount, startRadius, endRadius, startSize, endSize, startAngle, endAngel)"
+            yield "matrices = self.execute_Spiral(amount, startRadius, endRadius, startSize, endSize, startAngle, endAngel,\
+                                                  spiralHeight, radiusInterpolation, heightInterpolation)"
         elif self.mode == "SPLINE":
             if self.splineDistributionMethod == "STEP":
                 yield "matrices = self.execute_SplineStep(spline, step, start, end)"
@@ -254,7 +273,8 @@ class DistributeMatricesNode(bpy.types.Node, AnimationNode):
         return matrices
 
     def execute_Spiral(self, Py_ssize_t _amount, float startRadius, float endRadius,
-                             float startSize, float endSize, float startAngle, float endAngle):
+                             float startSize, float endSize, float startAngle, float endAngle,
+                             float spiralHeight, Interpolation radiusInterpolation, Interpolation heightInterpolation):
         cdef Py_ssize_t i
         cdef Vector3 position
         cdef float iCos, iSin, stepCos, stepSin, f, size
@@ -267,16 +287,18 @@ class DistributeMatricesNode(bpy.types.Node, AnimationNode):
         iSin = sin(startAngle)
         stepCos = cos(angleStep)
         stepSin = sin(angleStep)
-
+        
+        zOffset = (spiralHeight / 2) * self.centerSpiral
+        
         for i in range(amount):
             f = <float>i * factor
 
             size = f * (endSize - startSize) + startSize
-            radius = f * (endRadius - startRadius) + startRadius
+            radius = (endRadius - startRadius) * radiusInterpolation.evaluate(f) + startRadius
 
             position.x = iCos * radius
             position.y = iSin * radius
-            position.z = 0
+            position.z = spiralHeight * heightInterpolation.evaluate(f) - zOffset
 
             setTranslationMatrix(matrices.data + i, &position)
             setMatrixCustomZRotation(matrices.data + i, iCos, iSin)
@@ -289,10 +311,9 @@ class DistributeMatricesNode(bpy.types.Node, AnimationNode):
     def execute_SplineStep(self, Spline spline, float step, float start, float end):
         if not spline.isEvaluable(): return Matrix4x4List()
         spline.ensureUniformConverter(self.splineResolution)
-        start = spline.toUniformParameter(min(max(start, 0), 1))
-        end = spline.toUniformParameter(min(max(end, 0), 1))
-        spline.ensureNormals()
-        cdef float length = spline.getPartialLength(start, end, self.splineResolution)
+        cdef float uniformStart = spline.toUniformParameter(min(max(start, 0), 1))
+        cdef float uniformEnd = spline.toUniformParameter(min(max(end, 0), 1))
+        cdef float length = spline.getPartialLength(uniformStart, uniformEnd, self.splineResolution)
         cdef Py_ssize_t count = <Py_ssize_t>(length / step) if step != 0 else 0
         return self.execute_SplineCount(spline, count, start, end)
 
@@ -309,6 +330,7 @@ class DistributeMatricesNode(bpy.types.Node, AnimationNode):
 
     def execute_SplineVertices(self, Spline spline):
         if not spline.isEvaluable(): return Matrix4x4List()
+        spline.ensureNormals()
         count = len(spline.points)
         return spline.getDistributedMatrices(count, 0, 1, "RESOLUTION")